+++ AsahiApp/AsahiApp/CsPresentacion/CsPresentacion.csproj
@@ -93,10 +93,15 @@
       <DesignTime>True</DesignTime>
       <DependentUpon>asahi16DataSet.xsd</DependentUpon>
     </Compile>
-    <Compile Include="Ausentismo Global.cs">
+    <Compile Include="DS_Ausentismo_Global.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>DS_Ausentismo_Global.xsd</DependentUpon>
+    </Compile>
+    <Compile Include="Formularios\RH\Ausentismo Global.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Ausentismo Global.Designer.cs">
+    <Compile Include="Formularios\RH\Ausentismo Global.Designer.cs">
       <DependentUpon>Ausentismo Global.cs</DependentUpon>
     </Compile>
     <Compile Include="Detalle_incidencias.cs">
@@ -138,10 +143,10 @@
     <Compile Include="Formularios\RH\Permisos_Autorizaciones.Designer.cs">
       <DependentUpon>Permisos_Autorizaciones.cs</DependentUpon>
     </Compile>
-    <Compile Include="Frm_Rep_Historial.cs">
+    <Compile Include="Reportes\RH\Frm_Rep_Historial.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="Frm_Rep_Historial.Designer.cs">
+    <Compile Include="Reportes\RH\Frm_Rep_Historial.Designer.cs">
       <DependentUpon>Frm_Rep_Historial.cs</DependentUpon>
     </Compile>
     <Compile Include="Reportes\RH\Credencial_Visitantes.cs">
@@ -430,6 +435,12 @@
     <Compile Include="Reportes\RH\Rotacion_Grafico.Designer.cs">
       <DependentUpon>Rotacion_Grafico.cs</DependentUpon>
     </Compile>
+    <Compile Include="Reportes\RH\Rep_Ausentismo.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="Reportes\RH\Rep_Ausentismo.Designer.cs">
+      <DependentUpon>Rep_Ausentismo.cs</DependentUpon>
+    </Compile>
     <Compile Include="SqlServerTypes\Loader.cs" />
     <Compile Include="UserControl1.cs">
       <SubType>UserControl</SubType>
@@ -469,7 +480,8 @@
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Ausentismo Global.resx">
+    <EmbeddedResource Include="Ausentismo_Global.rdlc" />
+    <EmbeddedResource Include="Formularios\RH\Ausentismo Global.resx">
       <DependentUpon>Ausentismo Global.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Detalle_incidencias.resx">
@@ -484,7 +496,7 @@
     <EmbeddedResource Include="Formularios\RH\Permisos_Autorizaciones.resx">
       <DependentUpon>Permisos_Autorizaciones.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="Frm_Rep_Historial.resx">
+    <EmbeddedResource Include="Reportes\RH\Frm_Rep_Historial.resx">
       <DependentUpon>Frm_Rep_Historial.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Reportes\RH\Credencial_Visitantes.resx">
@@ -582,6 +594,9 @@
       <DependentUpon>Altas_Empleados.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Reporte_Incapacidades.rdlc" />
+    <EmbeddedResource Include="Reportes\RH\Rep_Ausentismo.resx">
+      <DependentUpon>Rep_Ausentismo.cs</DependentUpon>
+    </EmbeddedResource>
     <EmbeddedResource Include="Rep_Ausentismos.rdlc" />
     <EmbeddedResource Include="Rep_bajas.rdlc" />
     <EmbeddedResource Include="Reportes\RH\Finiquito.resx">
@@ -644,6 +659,17 @@
     </None>
     <None Include="DS_Ausentismos.xss">
       <DependentUpon>DS_Ausentismos.xsd</DependentUpon>
+    </None>
+    <None Include="DS_Ausentismo_Global.xsc">
+      <DependentUpon>DS_Ausentismo_Global.xsd</DependentUpon>
+    </None>
+    <None Include="DS_Ausentismo_Global.xsd">
+      <SubType>Designer</SubType>
+      <Generator>MSDataSetGenerator</Generator>
+      <LastGenOutput>DS_Ausentismo_Global.Designer.cs</LastGenOutput>
+    </None>
+    <None Include="DS_Ausentismo_Global.xss">
+      <DependentUpon>DS_Ausentismo_Global.xsd</DependentUpon>
     </None>
     <None Include="DS_Bajas.xsc">
       <DependentUpon>DS_Bajas.xsd</DependentUpon>
